+++ datalad/core/local/run.py
@@ -26,7 +26,7 @@ from tempfile import mkdtemp
 from datalad.core.local.save import Save
 from datalad.distribution.get import Get
 from datalad.distribution.install import Install
-from datalad.interface.unlock import Unlock
+from datalad.local.unlock import Unlock
 
 from datalad.interface.base import Interface
 from datalad.interface.utils import generic_result_renderer
