+++ datalad/interface/__init__.py
@@ -51,7 +51,7 @@ _group_dataset = (
         ('datalad.distributed.create_sibling_ria',
          'CreateSiblingRia',
          'create-sibling-ria'),
-        ('datalad.interface.unlock', 'Unlock', 'unlock'),
+        ('datalad.local.unlock', 'Unlock'),
         ('datalad.core.local.save', 'Save', 'save'),
         ('datalad.local.copy_file', 'CopyFile', 'copy-file'),
     ])
